Add files via upload
In this data there is some information about job id, title,location,salaryrange,company profile,description, requirements ,benefits etc. 
In this project we are going to classify that a news is real or fake . we will use one hot encoding for encoding the data and use column transformer to tranfer the columns , it will be done by Random Forest Classifier algorithm.